feat(auth): backend auth implementation
This commit adds the support for backend cookie session authentication inspired by xwehrenb.
Added endpoints are: login/register/get current user/logout.

The authorization supports roles, thus only a set of roles is allowed to interact
with the given endpoint if it is required.

Also added a insomnia export so anyone can test the routes.